Only show snippet delete button to owner or admin

+++ resources/views/snippets/show.blade.php
@@ -12,14 +12,15 @@
   <p>{{$snippet->description}}</p>
   <pre class="line-numbers code-view"><code class="language-{{$snippet->syntax}}">{{$snippet->text}}</code></pre>
 
-  {!! Form::open(['method'=>'delete', 'action'=>['SnippetsController@destroy', $snippet->id]]) !!}
-    {!! Form::submit('Delete Snippet', ['class'=>'btn btn-danger pull-right']) !!}
-  {!! Form::close() !!}
+  @if (Auth::user() && (Auth::user()->is_admin || Auth::id() == $snippet->user->id))
+    {!! Form::open(['method'=>'delete', 'action'=>['SnippetsController@destroy', $snippet->id]]) !!}
+      {!! Form::submit('Delete Snippet', ['class'=>'btn btn-danger pull-right']) !!}
+    {!! Form::close() !!}
+  @endif
 
   <a class="back-arrow" href="{{$back}}"><i class="fa fa-chevron-left"></i> Go Back</a>
 
   <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.6.0/prism.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.6.0/plugins/line-numbers/prism-line-numbers.js"></script>
   <script src="https://cdnjs.cloudflare.com/ajax/libs/prism/1.6.0/components/prism-{{$snippet->syntax}}.js"></script>
-
 @endsection
